+++ docs/zh/other/changelog.md
@@ -4,7 +4,7 @@ aside: false
 ---
 # 版本迭代记录
 
-## [4.0.0-SNAPSHOT](https://github.com/future0923/debug-tools/compare/v3.4.4...main)
+## [4.0.0](https://github.com/future0923/debug-tools/compare/v3.4.4...v4.0.0)(2025-06-03)
 
 - 增加热部署功能，实现秒级一键热更新远程应用代码。
 - 增加远程动态编译功能，热部署、热重载时可以选择本地Idea编译还是远程应用动态编译代码